requirements.txt
google-generativeai==0.8.3
numpy==1.26.4
pillow==11.0.0
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
opencv-python==4.10.0.84
python-dotenv==1.0.1
tensorflow==2.16.1


